Add lanuage selector, norwegian, franch and presistant language preference based on user
