+++ src/i18n/i18nChangeLanguage.js
@@ -1,0 +1,8 @@
+function ChangeLanguage(lang) {
+    window.locale = lang;
+    window.mutate_user_preferences({
+        language : lang,
+    });
+}
+
+export default ChangeLanguage;
