--- src/i18n/no.js
@@ -1,141 +1,0 @@
-const no = {
-    access_granted_to: "Tilgang gitt til",
-    add_existing_account: "Legg til eksisterende konto",
-    all_fields_required: "Alle felt er obligatoriske.",
-    apply: "Bruk",
-    ascending: "Stigende",
-    background: "Bakgrunn",
-    browse: "Bla gjennom",
-    cancel: "Avbryt",
-    center: "Sentrer",
-    change_desktop_background: "Endre skrivebordsbakgrunn…",
-    change_password: "Endre passord",
-    change_username: "Endre brukernavn",
-    close_all_windows: "Lukk alle vinduer",
-    color: "Farge",
-    confirm_account_for_free_referral_storage_c2a: "Opprett en konto og bekreft e-postadressen din for å motta 1 GB gratis lagringsplass. Din venn vil også få 1 GB gratis lagringsplass.",
-    confirm_new_password: "Bekreft nytt passord",
-    contact_us: "Kontakt oss",
-    contain: "Inneholde",
-    continue: "Fortsett",
-    copy: "Kopier",
-    copy_link: "Kopier lenke",
-    copying: "Kopierer",
-    cover: "Dekke",
-    create_account: "Opprett konto",
-    create_free_account: "Opprett gratis konto",
-    create_shortcut: "Opprett snarvei",
-    current_password: "Nåværende passord",
-    cut: "Klipp ut",
-    date_modified: "Endret dato",
-    delete: "Slett",
-    delete_permanently: "Slett permanent",
-    deploy_as_app: "Distribuer som app",
-    descending: "Synkende",
-    desktop_background_fit: "Tilpass",
-    dir_published_as_website: "%strong% er publisert på:",
-    disassociate_dir: "Fjern tilknytning fra mappe",
-    download: "Last ned",
-    downloading: "Laster ned",
-    email: "E-post",
-    email_or_username: "E-post eller brukernavn",
-    empty_trash: "Tøm papirkurv",
-    empty_trash_confirmation: "Er du sikker på at du vil slette alt i papirkurven permanent?",
-    emptying_trash: "Tømmer papirkurv…",
-    feedback: "Tilbakemelding",
-    feedback_c2a: "Vennligst bruk skjemaet nedenfor for å sende oss din tilbakemelding, kommentarer og feilrapporter.",
-    feedback_sent_confirmation: "Takk for at du kontaktet oss. Hvis du har en e-post knyttet til kontoen din, vil du høre fra oss så snart som mulig.",
-    forgot_pass_c2a: "Glemt passord?",
-    from: "Fra",
-    general: "Generelt",
-    get_a_copy_of_on_puter: "Få en kopi av '%%' på Puter.com!",
-    get_copy_link: "Få kopilenke",
-    hide_all_windows: "Skjul alle vinduer",
-    html_document: "HTML-dokument",
-    image: "Bilde",
-    invite_link: "Invitasjonslenke",
-    items_in_trash_cannot_be_renamed: "Dette elementet kan ikke omdøpes fordi det er i papirkurven. For å omdøpe dette elementet, dra det først ut av papirkurven.",
-    jpeg_image: "JPEG-bilde",
-    keep_in_taskbar: "Behold i oppgavelinjen",
-    log_in: "Logg inn",
-    log_out: "Logg ut",
-    move: "Flytt",
-    moving: "Flytter",
-    my_websites: "Mine nettsteder",
-    name: "Navn",
-    name_cannot_be_empty: "Navn kan ikke være tomt.",
-    name_cannot_contain_double_period: "Navn kan ikke inneholde '..'.",
-    name_cannot_contain_period: "Navn kan ikke inneholde '.'-tegnet.",
-    name_cannot_contain_slash: "Navn kan ikke inneholde '/'-tegnet.",
-    name_must_be_string: "Navn kan bare være en streng.",
-    name_too_long: "Navn kan ikke være lengre enn %% tegn.",
-    new: "Ny",
-    new_folder: "Ny mappe",
-    new_password: "Nytt passord",
-    new_username: "Nytt brukernavn",
-    no_dir_associated_with_site: "Ingen mappe er tilknyttet denne adressen.",
-    no_websites_published: "Du har ikke publisert noen nettsteder ennå.",
-    ok: "OK",
-    open: "Åpne",
-    open_in_new_tab: "Åpne i ny fane",
-    open_in_new_window: "Åpne i nytt vindu",
-    open_with: "Åpne med",
-    password: "Passord",
-    password_changed: "Passord endret.",
-    passwords_do_not_match: "`Nytt passord` og `Bekreft nytt passord` stemmer ikke overens.",
-    paste: "Lim inn",
-    paste_into_folder: "Lim inn i mappe",
-    pick_name_for_website: "Velg et navn for nettstedet ditt:",
-    picture: "Bilde",
-    powered_by_puter_js: "Drevet av <a href=\"https://docs.puter.com/\" target=\"_blank\">Puter.js</a>",
-    preparing: "Forbereder...",
-    preparing_for_upload: "Forbereder opplasting...",
-    properties: "Egenskaper",
-    publish: "Publiser",
-    publish_as_website: "Publiser som nettsted",
-    recent: "Nylig",
-    recover_password: "Gjenopprett passord",
-    refer_friends_c2a: "Få 1 GB for hver venn som oppretter og bekrefter en konto på Puter. Vennen din får også 1 GB.",
-    refer_friends_social_media_c2a: "Få 1 GB gratis lagringsplass på Puter.com!",
-    refresh: "Oppdater",
-    release_address_confirmation: "Er du sikker på at du vil frigi denne adressen?",
-    remove_from_taskbar: "Fjern fra oppgavelinjen",
-    rename: "Gi nytt navn",
-    repeat: "Gjenta",
-    resend_confirmation_code: "Send bekreftelseskoden på nytt",
-    restore: "Gjenopprett",
-    save_account_to_get_copy_link: "Vennligst opprett en konto for å fortsette.",
-    save_account_to_publish: "Vennligst opprett en konto for å fortsette.",
-    save_session_c2a: "Opprett en konto for å lagre gjeldende økt og unngå å miste arbeidet ditt.",
-    scan_qr_c2a: "Skann koden nedenfor for å logge inn på denne økten fra andre enheter",
-    select: "Velg",
-    select_color: "Velg farge…",
-    send: "Send",
-    send_password_recovery_email: "Send e-post for gjenoppretting av passord",
-    session_saved: "Takk for at du opprettet en konto. Denne økten er lagret.",
-    set_new_password: "Angi nytt passord",
-    share_to: "Del",
-    show_all_windows: "Vis alle vinduer",
-    show_hidden: "Vis skjulte",
-    sign_in_with_puter: "Logg inn med Puter",
-    sign_up: "Registrer deg",
-    signing_in: "Logger inn…",
-    size: "Størrelse",
-    sort_by: "Sorter etter",
-    start: "Start",
-    taking_longer_than_usual: "Dette tar litt lenger tid enn vanlig. Vennligst vent...",
-    text_document: "Tekstdokument",
-    tos_fineprint: "Ved å klikke på 'Opprett gratis konto' godtar du Puters <a href=\"https://puter.com/terms\" target=\"_blank\">tjenestevilkår</a> og <a href=\"https://puter.com/privacy\" target=\"_blank\">personvernpolicy</a>.",
-    trash: "Papirkurv",
-    type: "Type",
-    undo: "Angre",
-    unzip: "Pakk ut",
-    upload: "Last opp",
-    upload_here: "Last opp her",
-    username: "Brukernavn",
-    username_changed: "Brukernavn oppdatert.",
-    versions: "Versjoner",
-    yes_release_it: "Ja, frigi den",
-    you_have_been_referred_to_puter_by_a_friend: "Du har blitt henvist til Puter av en venn!",
-    zip: "Zip"
-};
