+++ src/i18n/translations/en.js
@@ -1,0 +1,148 @@
+const en = {
+    name: "English",
+    code: "en",
+    dictionary: {
+        access_granted_to: "Access Granted To",
+        add_existing_account: "Add Existing Account",
+        all_fields_required: 'All fields are required.',
+        apply: "Apply",
+        ascending: 'Ascending',
+        background: "Background",
+        browse: "Browse",
+        cancel: 'Cancel',
+        center: 'Center',
+        change_desktop_background: 'Change desktop background…',
+        change_language: "Change Language",
+        change_password: "Change Password",
+        change_username: "Change Username",
+        close_all_windows: "Close All Windows",
+        color: 'Color',
+        confirm_account_for_free_referral_storage_c2a: 'Create an account and confirm your email address to receive 1 GB of free storage. Your friend will get 1 GB of free storage too.',
+        confirm_new_password: "Confirm New Password",
+        contact_us: "Contact Us",
+        contain: 'Contain',
+        continue: "Continue",
+        copy: 'Copy',
+        copy_link: "Copy Link",
+        copying: "Copying",
+        cover: 'Cover',
+        create_account: "Create Account",
+        create_free_account: "Create Free Account",
+        create_shortcut: "Create Shortcut",
+        current_password: "Current Password",
+        cut: 'Cut',
+        date_modified: 'Date modified',
+        delete: 'Delete',
+        delete_permanently: "Delete Permanently",
+        deploy_as_app: 'Deploy as app',
+        descending: 'Descending',
+        desktop_background_fit: "Fit",
+        dir_published_as_website: `%strong% has been published to:`,
+        disassociate_dir: "Disassociate Directory",
+        download: 'Download',
+        downloading: "Downloading",
+        email: "Email",
+        email_or_username: "Email or Username",
+        empty_trash: 'Empty Trash',
+        empty_trash_confirmation: `Are you sure you want to permanently delete the items in Trash?`,
+        emptying_trash: 'Emptying Trash…',
+        feedback: "Feedback",
+        feedback_c2a: "Please use the form below to send us your feedback, comments, and bug reports.",
+        feedback_sent_confirmation: "Thank you for contacting us. If you have an email associated with your account, you will hear back from us as soon as possible.",
+        forgot_pass_c2a: "Forgot password?",
+        from: "From",
+        general: "General",
+        get_a_copy_of_on_puter: `Get a copy of '%%' on Puter.com!`,
+        get_copy_link: 'Get Copy Link',
+        hide_all_windows: "Hide All Windows",
+        html_document: 'HTML document',
+        image: 'Image',
+        invite_link: "Invite Link",
+        items_in_trash_cannot_be_renamed: `This item can't be renamed because it's in the trash. To rename this item, first drag it out of the Trash.`,
+        jpeg_image: 'JPEG image',
+        keep_in_taskbar: 'Keep in Taskbar',
+        log_in: "Log In",
+        log_out: 'Log Out',
+        move: 'Move',
+        moving: "Moving",
+        my_websites: "My Websites",
+        name: 'Name',
+        name_cannot_be_empty: 'Name cannot be empty.',
+        name_cannot_contain_double_period: "Name can not be the '..' character.",
+        name_cannot_contain_period: "Name can not be the '.' character.",
+        name_cannot_contain_slash: "Name cannot contain the '/' character.",
+        name_must_be_string: "Name can only be a string.",
+        name_too_long: `Name can not be longer than %% characters.`,
+        new: 'New',
+        new_folder: 'New folder',
+        new_password: "New Password",
+        new_username: "New Username",
+        no_dir_associated_with_site: 'No directory associated with this address.',
+        no_websites_published: "You have not published any websites yet.",
+        ok: 'OK',
+        open: "Open",
+        open_in_new_tab: "Open in New Tab",
+        open_in_new_window: "Open in New Window",
+        open_with: "Open With",
+        password: "Password",
+        password_changed: "Password changed.",
+        passwords_do_not_match: '`New Password` and `Confirm New Password` do not match.',
+        paste: 'Paste',
+        paste_into_folder: "Paste Into Folder",
+        pick_name_for_website: "Pick a name for your website:",
+        picture: "Picture",
+        powered_by_puter_js: `Powered by <a href="https://docs.puter.com/" target="_blank">Puter.js</a>`,
+        preparing: "Preparing...",
+        preparing_for_upload: "Preparing for upload...",
+        properties: "Properties",
+        publish: "Publish",
+        publish_as_website: 'Publish as website',
+        recent: "Recent",
+        recover_password: "Recover Password",
+        refer_friends_c2a: "Get 1 GB for every friend who creates and confirms an account on Puter. Your friend will get 1 GB too!",
+        refer_friends_social_media_c2a: `Get 1 GB of free storage on Puter.com!`,
+        refresh: 'Refresh',
+        release_address_confirmation: `Are you sure you want to release this address?`,
+        remove_from_taskbar:'Remove from Taskbar',
+        rename: 'Rename',
+        repeat: 'Repeat',
+        resend_confirmation_code: "Re-send Confirmation Code",
+        restore: "Restore",
+        save_account_to_get_copy_link: "Please create an account to proceed.",
+        save_account_to_publish: 'Please create an account to proceed.',
+        save_session_c2a: 'Create an account to save your current session and avoid losing your work.',
+        scan_qr_c2a: 'Scan the code below to log into this session from other devices',
+        select: "Select",
+        select_color: 'Select color…',
+        send: "Send",
+        send_password_recovery_email: "Send Password Recovery Email",
+        session_saved: "Thank you for creating an account. This session has been saved.",
+        set_new_password: "Set New Password",
+        share_to: "Share to",
+        show_all_windows: "Show All Windows",
+        show_hidden: 'Show hidden',
+        sign_in_with_puter: "Sign in with Puter",
+        sign_up: "Sign Up",
+        signing_in: "Signing in…",
+        size: 'Size',
+        sort_by: 'Sort by',
+        start: 'Start',
+        taking_longer_than_usual: 'Taking a little longer than usual. Please wait...',
+        text_document: 'Text document',
+        tos_fineprint: `By clicking 'Create Free Account' you agree to Puter's <a href="https://puter.com/terms" target="_blank">Terms of Service</a> and <a href="https://puter.com/privacy" target="_blank">Privacy Policy</a>.`,
+        trash: 'Trash',
+        type: 'Type',
+        undo: 'Undo',
+        unzip: "Unzip",
+        upload: 'Upload',
+        upload_here: 'Upload here',
+        username: "Username",
+        username_changed: 'Username updated successfully.',
+        versions: "Versions",
+        yes_release_it: 'Yes, Release It',
+        you_have_been_referred_to_puter_by_a_friend: "You have been referred to Puter by a friend!",
+        zip: "Zip",
+    }
+};
+
+export default en;
