+++ src/i18n/translations/fa.js
@@ -1,0 +1,148 @@
+const fa = {
+    name: "فارسی",
+    code: "fa",
+    dictionary: {
+        access_granted_to: "دسترسی داده شده به",
+        add_existing_account: "افزودن حساب کاربری موجود",
+        all_fields_required: 'تمامی فیلدها الزامی هستند.',
+        apply: "اعمال",
+        ascending: 'صعودی',
+        background: "پس زمینه",
+        browse: "مرور",
+        cancel: 'لغو',
+        center: 'مرکز',
+        change_desktop_background: 'تغییر پس زمینه دسکتاپ…',
+        change_language: "تغییر زبان",
+        change_password: "تغییر رمز عبور",
+        change_username: "تغییر نام کاربری",
+        close_all_windows: "بستن همه پنجره ها",
+        color: 'رنگ',
+        confirm_account_for_free_referral_storage_c2a: 'حساب کاربری خود را ایجاد کرده و آدرس ایمیل خود را تأیید کنید تا 1 گیگابایت فضای ذخیره سازی رایگان دریافت کنید. دوست شما هم 1 گیگابایت فضای ذخیره سازی رایگان دریافت خواهد کرد.',
+        confirm_new_password: "تأیید رمز عبور جدید",
+        contact_us: "تماس با ما",
+        contain: 'شامل',
+        continue: "ادامه",
+        copy: 'کپی',
+        copy_link: "کپی لینک",
+        copying: "کپی",
+        cover: 'جلد',
+        create_account: "ایجاد حساب کاربری",
+        create_free_account: "ایجاد حساب کاربری رایگان",
+        create_shortcut: "ایجاد میانبر",
+        current_password: "رمز عبور فعلی",
+        cut: 'برش',
+        date_modified: 'تاریخ تغییر',
+        delete: 'حذف',
+        delete_permanently: "حذف دائمی",
+        deploy_as_app: 'نصب به عنوان برنامه',
+        descending: 'نزولی',
+        desktop_background_fit: "متناسب",
+        dir_published_as_website: `%strong% منتشر شده به:`,
+        disassociate_dir: "قطع ارتباط دایرکتوری",
+        download: 'دانلود',
+        downloading: "دانلود",
+        email: "ایمیل",
+        email_or_username: "ایمیل یا نام کاربری",
+        empty_trash: 'خالی کردن سطل زباله',
+        empty_trash_confirmation: `آیا از حذف دائمی موارد در سطل زباله مطمئن هستید؟`,
+        emptying_trash: 'خالی کردن سطل زباله…',
+        feedback: "بازخورد",
+        feedback_c2a: "لطفا از فرم زیر برای ارسال بازخورد، نظرات و گزارش خطا استفاده کنید.",
+        feedback_sent_confirmation: "با تشکر از تماس شما. اگر ایمیلی به حساب کاربری شما متصل است، در اسرع وقت پاسخ خواهیم داد.",
+        forgot_pass_c2a: "رمز عبور را فراموش کرده اید؟",
+        from: "از",
+        general: "عمومی",
+        get_a_copy_of_on_puter: `یک نسخه از '%%' را در Puter.com بگیرید!`,
+        get_copy_link: 'گرفتن لینک کپی',
+        hide_all_windows: "پنهان کردن همه پنجره ها",
+        html_document: 'سند HTML',
+        image: 'تصویر',
+        invite_link: "لینک دعوت",
+        items_in_trash_cannot_be_renamed: `این مورد نمی تواند تغییر نام دهد زیرا در سطل زباله است. برای تغییر نام این مورد، ابتدا آن را از سطل زباله بیرون بکشید.`,
+        jpeg_image: 'تصویر JPEG',
+        keep_in_taskbar: 'در نوار وظایف نگه دارید',
+        log_in: "ورود",
+        log_out: 'خروج',
+        move: 'انتقال',
+        moving: "انتقال",
+        my_websites: "وبسایت های من",
+        name: 'نام',
+        name_cannot_be_empty: 'نام نمی تواند خالی باشد.',
+        name_cannot_contain_double_period: "نام نمی تواند شامل '..' باشد.",
+        name_cannot_contain_period: "نام نمی تواند شامل '.' باشد.",
+        name_cannot_contain_slash: "نام نمی تواند شامل '/' باشد.",
+        name_must_be_string: "نام فقط می تواند یک رشته باشد.",
+        name_too_long: `نام نمی تواند بیشتر از %% کاراکتر باشد.`,
+        new: 'جدید',
+        new_folder: 'پوشه جدید',
+        new_password: "رمز عبور جدید",
+        new_username: "نام کاربری جدید",
+        no_dir_associated_with_site: 'هیچ دایرکتوری مرتبط با این آدرس وجود ندارد.',
+        no_websites_published: "هنوز هیچ وبسایتی منتشر نکرده اید.",
+        ok: 'خوب',
+        open: "باز کردن",
+        open_in_new_tab: "در تب جدید باز کن",
+        open_in_new_window: "در پنجره جدید باز کن",
+        open_with: "باز کردن با",
+        password: "رمز عبور",
+        password_changed: "رمز عبور تغییر یافت.",
+        passwords_do_not_match: '`رمز عبور جدید` و `تأیید رمز عبور جدید` مطابقت ندارند.',
+        paste: 'چسباندن',
+        paste_into_folder: "چسباندن در پوشه",
+        pick_name_for_website: "یک نام برای وبسایت خود انتخاب کنید:",
+        picture: "تصویر",
+        powered_by_puter_js: `پشتیبانی شده توسط <a href="https://docs.puter.com/" target="_blank">Puter.js</a>`,
+        preparing: "در حال آماده سازی...",
+        preparing_for_upload: "آماده سازی برای بارگذاری...",
+        properties: "ویژگی ها",
+        publish: "انتشار",
+        publish_as_website: 'انتشار به عنوان وبسایت',
+        recent: "اخیر",
+        recover_password: "بازیابی رمز عبور",
+        refer_friends_c2a: "برای هر دوستی که حساب کاربری Puter ایجاد و تأیید کند، 1 گیگابایت دریافت کنید. دوست شما هم 1 گیگابایت دریافت خواهد کرد!",
+        refer_friends_social_media_c2a: `1 گیگابایت فضای ذخیره سازی رایگان را در Puter.com بگیرید!`,
+        refresh: 'تازه کردن',
+        release_address_confirmation: `آیا مطمئن هستید که می خواهید این آدرس را آزاد کنید؟`,
+        remove_from_taskbar:'از نوار وظایف حذف کن',
+        rename: 'تغییر نام',
+        repeat: 'تکرار',
+        resend_confirmation_code: "ارسال مجدد کد تأیید",
+        restore: "بازیابی",
+        save_account_to_get_copy_link: "لطفا برای ادامه یک حساب کاربری ایجاد کنید.",
+        save_account_to_publish: 'لطفا برای ادامه یک حساب کاربری ایجاد کنید.',
+        save_session_c2a: 'برای ذخیره جلسه فعلی و جلوگیری از از دست دادن کار خود یک حساب کاربری ایجاد کنید.',
+        scan_qr_c2a: 'کد زیر را از دستگاه های دیگر اسکن کنید تا به این جلسه وارد شوید',
+        select: "انتخاب",
+        select_color: 'انتخاب رنگ…',
+        send: "ارسال",
+        send_password_recovery_email: "ارسال ایمیل بازیابی رمز عبور",
+        session_saved: "با تشکر از ایجاد حساب کاربری. این جلسه ذخیره شده است.",
+        set_new_password: "تنظیم رمز عبور جدید",
+        share_to: "اشتراک گذاری به",
+        show_all_windows: "نمایش همه پنجره ها",
+        show_hidden: 'نمایش مخفی',
+        sign_in_with_puter: "ورود با Puter",
+        sign_up: "ثبت نام",
+        signing_in: "ورود…",
+        size: 'اندازه',
+        sort_by: 'مرتب سازی بر اساس',
+        start: 'شروع',
+        taking_longer_than_usual: 'کمی بیشتر از معمول طول می کشد. لطفا صبر کنید...',
+        text_document: 'سند متنی',
+        tos_fineprint: `با کلیک بر روی 'ایجاد حساب کاربری رایگان' شما با <a href="https://puter.com/terms" target="_blank">شرایط خدمات</a> و <a href="https://puter.com/privacy" target="_blank">سیاست حفظ حریم خصوصی</a> Puter موافقت می کنید.`,
+        trash: 'سطل زباله',
+        type: 'نوع',
+        undo: 'بازگشت',
+        unzip: "باز کردن فایل فشرده",
+        upload: 'بارگذاری',
+        upload_here: 'اینجا بارگذاری کنید',
+        username: "نام کاربری",
+        username_changed: 'نام کاربری با موفقیت به روز شد.',
+        versions: "نسخه ها",
+        yes_release_it: 'بله، آن را آزاد کن',
+        you_have_been_referred_to_puter_by_a_friend: "شما توسط یک دوست به Puter معرفی شده اید!",
+        zip: "فشرده سازی",
+    }
+};
+
+export default fa;
