+++ src/i18n/translations/fr.js
@@ -1,0 +1,148 @@
+const fr = {
+    name: "Français",
+    code: "fr",
+    dictionary: {
+        access_granted_to: "Accès accordé à",
+        add_existing_account: "Ajouter un compte existant",
+        all_fields_required: 'All fields are required.',
+        apply: "Appliquer",
+        ascending: 'Ascending',
+        background: "Background",
+        browse: "Browse",
+        cancel: 'Annuler',
+        center: 'Center',
+        change_desktop_background: 'Changer l’arrière-plan du bureau…',
+        change_language: "Changer de langue",
+        change_password: "Changer de mot de passe",
+        change_username: "Changer de nom d'ulilisateur",
+        close_all_windows: "Fermer toutes les fenêtres",
+        color: 'Couleur',
+        confirm_account_for_free_referral_storage_c2a: 'Créez un compte et confirmez votre adresse e-mail pour recevoir 1 Go de stockage gratuit. Votre ami bénéficiera également de 1 Go de stockage gratuit.',
+        confirm_new_password: "Confirmez le nouveau mot de passe",
+        contact_us: "Nous contacter",
+        contain: 'Contain',
+        continue: "Continuer",
+        copy: 'Copier',
+        copy_link: "Copier le lien",
+        copying: "Copying",
+        cover: 'Cover',
+        create_account: "Créer un compte",
+        create_free_account: "Créer un compte gratuit",
+        create_shortcut: "Créer un raccourci",
+        current_password: "Mot de passe actuel",
+        cut: 'Couper',
+        date_modified: 'Date de modification',
+        delete: 'Supprimer',
+        delete_permanently: "Supprimer définitivement",
+        deploy_as_app: 'Deployer en tant qu\'application',
+        descending: 'Décroissant',
+        desktop_background_fit: "Ajuster",
+        dir_published_as_website: `%strong% a été publié sur :`,
+        disassociate_dir: "Dissocier le répertoire",
+        download: 'Télécharger',
+        downloading: "Téléchargement en cours",
+        email: "Email",
+        email_or_username: "Email ou nom d'utilisateur",
+        empty_trash: 'Vider la corbeille',
+        empty_trash_confirmation: `Êtes-vous sûr de vouloir supprimer définitivement les éléments de la corbeille ?`,
+        emptying_trash: 'Vidage de la corbeille...',
+        feedback: "Envoyer des commentaires",
+        feedback_c2a: "Veuillez utiliser le formulaire ci-dessous pour nous envoyer vos retours, commentaires et rapports de bugs.",
+        feedback_sent_confirmation: "Merci de nous avoir contactés. Si vous avez un e-mail associé à votre compte, vous recevrez une réponse de notre part dans les plus brefs délais.",
+        forgot_pass_c2a: "Mot de passe oublier?",
+        from: "De",
+        general: "Général",
+        get_a_copy_of_on_puter: `Obtenez une copie de '%%' sur Puter.com!`,
+        get_copy_link: 'Obtenir le lien de copie',
+        hide_all_windows: "Masquer toutes les fenêtres",
+        html_document: 'Document HTML',
+        image: 'Image',
+        invite_link: "Lien d'invitation",
+        items_in_trash_cannot_be_renamed: `Cet élément ne peut pas être renommé car il se trouve dans la corbeille. Pour renommer cet élément, faites-le d'abord glisser hors de la corbeille.`,
+        jpeg_image: 'Image JPEG',
+        keep_in_taskbar: 'Garder dans la barre des tâches',
+        log_in: "Se connecter",
+        log_out: 'Déconnexion',
+        move: 'Déplacer',
+        moving: "Déplacement en cours",
+        my_websites: "Mes sites internet",
+        name: 'Nom',
+        name_cannot_be_empty: 'Le nom ne peut pas être vide.',
+        name_cannot_contain_double_period: "Le nom ne peut pas être le caractère '..'.",
+        name_cannot_contain_period: "Le nom ne peut pas être le caractère '.'.",
+        name_cannot_contain_slash: "Le nom ne peut pas contenir le caractère '/'.",
+        name_must_be_string: "Le nom peut uniquement être une chaîne.",
+        name_too_long: `Le nom ne peut pas contenir plus de %% caractères.`,
+        new: 'Nouveau',
+        new_folder: 'Nouveau dossier',
+        new_password: "Nouveau mot de passe",
+        new_username: "Nouveau nom d'utilisateur",
+        no_dir_associated_with_site: 'Aucun répertoire associé à cette adresse.',
+        no_websites_published: "Vous n'avez pas encore publié de sites internet.",
+        ok: 'OK',
+        open: "Ouvrir",
+        open_in_new_tab: "Ouvrir dans un nouvel onglet",
+        open_in_new_window: "Ouvrir dans une nouvelle fenêtre",
+        open_with: "Ouvrir avec",
+        password: "Mot de passe",
+        password_changed: "Mot de passe modifié.",
+        passwords_do_not_match: '\'Nouveau mot de passe\' et \'Confirmer le nouveau mot de passe\' ne correspondent pas.',
+        paste: 'Coller',
+        paste_into_folder: "Coller dans le dossier",
+        pick_name_for_website: "Choisissez un nom pour votre site Web :",
+        picture: "Image",
+        powered_by_puter_js: `Alimenté par <a href="https://docs.puter.com/" target="_blank">Puter.js</a>`,
+        preparing: "Préparation...",
+        preparing_for_upload: "Préparation du chargement...",
+        properties: "Propriétés",
+        publish: "Publier",
+        publish_as_website: 'Publier en tant que site internet',
+        recent: "Récent",
+        recover_password: "Récupérer le mot de passe",
+        refer_friends_c2a: "Obtenez 1 Go pour chaque ami qui crée et confirme un compte sur Puter. Votre ami recevra également 1 Go !",
+        refer_friends_social_media_c2a: `Obtenez 1 Go de stockage gratuit sur Puter.com !`,
+        refresh: 'Actualiser',
+        release_address_confirmation: `Êtes-vous sûr de vouloir libérer cette adresse ?`,
+        remove_from_taskbar:'Retirer de la barre des tâches',
+        rename: 'Renomer',
+        repeat: 'Répéter',
+        resend_confirmation_code: "Renvoyer le code de confirmation",
+        restore: "Restorer",
+        save_account_to_get_copy_link: "Veuillez créer un compte pour continuer",
+        save_account_to_publish: 'Veuillez créer un compte pour continuer.',
+        save_session_c2a: 'Créez un compte pour sauvegarder votre session en cours et éviter de perdre votre travail.',
+        scan_qr_c2a: 'Scannez le code ci-dessous pour vous connecter à cette session depuis d\'autres appareils',
+        select: "Sélectionner",
+        select_color: 'Sélectionner une couleur...',
+        send: "Envoyer",
+        send_password_recovery_email: "Envoyer un e-mail de récupération de mot de passe",
+        session_saved: "Merci d'avoir créé un compte. Cette session a été sauvegardée.",
+        set_new_password: "Definir un nouveau mot de passe",
+        share_to: "Partager à",
+        show_all_windows: "Afficher toutes les fenêtres",
+        show_hidden: 'Afficher les éléments cachés',
+        sign_in_with_puter: "Se connecter avec Puter",
+        sign_up: "S'inscrire",
+        signing_in: "Connexion…",
+        size: 'Taille',
+        sort_by: 'Trier par',
+        start: 'Démarrer',
+        taking_longer_than_usual: 'Cela prend un peu plus de temps que d\'habitude. Veuillez patienter...',
+        text_document: 'Document texte',
+        tos_fineprint: `En cliquant sur "Créer un compte gratuit", vous acceptez les <a href="https://puter.com/terms" target="_blank">conditions d'utilisation</a> et la <a href="https://puter .com/privacy" target="_blank">politique de confidentialité</a>.`,
+        trash: 'Corbeille',
+        type: 'Type',
+        undo: 'Annuler',
+        unzip: "Décompresser",
+        upload: 'Charger',
+        upload_here: 'Charger ici',
+        username: "Nom d'utilisateur",
+        username_changed: 'Nom d\'utilisateur mis à jour avec succès.',
+        versions: "Versions",
+        yes_release_it: 'Oui, libérez-la',
+        you_have_been_referred_to_puter_by_a_friend: "Vous avez été référé à Puter par un ami !",
+        zip: "Compresser",
+    }
+}
+
+export default fr;
