+++ src/i18n/translations/ko.js
@@ -1,0 +1,148 @@
+const ko = {
+    name: "한국어",
+    code: "ko",
+    dictionary: {
+        access_granted_to: "접근 권한 부여",
+        add_existing_account: "기존 계정 추가",
+        all_fields_required: '모든 필드는 필수입니다.',
+        apply: "적용",
+        ascending: '오름차순',
+        background: "배경",
+        browse: "찾아보기",
+        cancel: '취소',
+        center: '중앙',
+        change_desktop_background: '바탕 화면 배경 변경…',
+        change_language: "언어 변경",
+        change_password: "비밀번호 변경",
+        change_username: "사용자 이름 변경",
+        close_all_windows: "모든 창 닫기",
+        color: '색상',
+        confirm_account_for_free_referral_storage_c2a: '계정을 생성하고 이메일 주소를 확인하여 1GB의 무료 저장 공간을 받으십시오. 친구도 1GB의 무료 저장 공간을 받게 됩니다.',
+        confirm_new_password: "새 비밀번호 확인",
+        contact_us: "문의하기",
+        contain: '포함',
+        continue: "계속",
+        copy: '복사',
+        copy_link: "링크 복사",
+        copying: "복사 중",
+        cover: '표지',
+        create_account: "계정 생성",
+        create_free_account: "무료 계정 생성",
+        create_shortcut: "바로 가기 만들기",
+        current_password: "현재 비밀번호",
+        cut: '잘라내기',
+        date_modified: '수정한 날짜',
+        delete: '삭제',
+        delete_permanently: "영구 삭제",
+        deploy_as_app: '앱으로 배포',
+        descending: '내림차순',
+        desktop_background_fit: "맞추기",
+        dir_published_as_website: `%strong% 다음에 게시되었습니다:`,
+        disassociate_dir: "디렉토리 연결 해제",
+        download: '다운로드',
+        downloading: "다운로드 중",
+        email: "이메일",
+        email_or_username: "이메일 또는 사용자 이름",
+        empty_trash: '휴지통 비우기',
+        empty_trash_confirmation: `휴지통의 항목을 영구적으로 삭제하시겠습니까?`,
+        emptying_trash: '휴지통 비우는 중…',
+        feedback: "피드백",
+        feedback_c2a: "아래 양식을 사용하여 피드백, 의견 및 버그 보고를 보내십시오.",
+        feedback_sent_confirmation: "문의해 주셔서 감사합니다. 계정에 이메일이 연결되어 있으면 가능한 빨리 회신 드리겠습니다.",
+        forgot_pass_c2a: "비밀번호를 잊으셨나요?",
+        from: "보낸 사람",
+        general: "일반",
+        get_a_copy_of_on_puter: `Puter.com에서 '%%'의 사본을 받으세요!`,
+        get_copy_link: '링크 복사',
+        hide_all_windows: "모든 창 숨기기",
+        html_document: 'HTML 문서',
+        image: '이미지',
+        invite_link: "초대 링크",
+        items_in_trash_cannot_be_renamed: `이 항목은 휴지통에 있기 때문에 이름을 바꿀 수 없습니다. 이 항목의 이름을 바꾸려면 먼저 휴지통에서 끌어내십시오.`,
+        jpeg_image: 'JPEG 이미지',
+        keep_in_taskbar: '작업 표시줄에 유지',
+        log_in: "로그인",
+        log_out: '로그아웃',
+        move: '이동',
+        moving: "이동 중",
+        my_websites: "내 웹사이트",
+        name: '이름',
+        name_cannot_be_empty: '이름은 비워둘 수 없습니다.',
+        name_cannot_contain_double_period: "이름은 '..' 문자일 수 없습니다.",
+        name_cannot_contain_period: "이름은 '.' 문자일 수 없습니다.",
+        name_cannot_contain_slash: "이름에 '/' 문자를 포함할 수 없습니다.",
+        name_must_be_string: "이름은 문자열만 가능합니다.",
+        name_too_long: `이름은 %%자보다 길 수 없습니다.`,
+        new: '새로운',
+        new_folder: '새 폴더',
+        new_password: "새 비밀번호",
+        new_username: "새 사용자 이름",
+        no_dir_associated_with_site: '이 주소에 연결된 디렉토리가 없습니다.',
+        no_websites_published: "아직 웹사이트를 게시하지 않았습니다.",
+        ok: '확인',
+        open: "열기",
+        open_in_new_tab: "새 탭에서 열기",
+        open_in_new_window: "새 창에서 열기",
+        open_with: "열기 방법",
+        password: "비밀번호",
+        password_changed: "비밀번호가 변경되었습니다.",
+        passwords_do_not_match: '`새 비밀번호`와 `새 비밀번호 확인`이 일치하지 않습니다.',
+        paste: '붙여넣기',
+        paste_into_folder: "폴더에 붙여넣기",
+        pick_name_for_website: "웹사이트 이름을 선택하세요:",
+        picture: "사진",
+        powered_by_puter_js: `Powered by <a href="https://docs.puter.com/" target="_blank">Puter.js</a>`,
+        preparing: "준비 중...",
+        preparing_for_upload: "업로드 준비 중...",
+        properties: "속성",
+        publish: "게시",
+        publish_as_website: '웹사이트로 게시',
+        recent: "최근",
+        recover_password: "비밀번호 찾기",
+        refer_friends_c2a: "Puter에서 계정을 생성하고 확인한 친구마다 1GB를 받으십시오. 친구도 1GB를 받게 됩니다!",
+        refer_friends_social_media_c2a: `Puter.com에서 1GB의 무료 저장 공간을 받으십시오!`,
+        refresh: '새로 고침',
+        release_address_confirmation: `이 주소를 해제하시겠습니까?`,
+        remove_from_taskbar:'작업 표시줄에서 제거',
+        rename: '이름 바꾸기',
+        repeat: '반복',
+        resend_confirmation_code: "확인 코드 다시 보내기",
+        restore: "복원",
+        save_account_to_get_copy_link: "계속하려면 계정을 생성하십시오.",
+        save_account_to_publish: '계속하려면 계정을 생성하십시오.',
+        save_session_c2a: '현재 세션을 저장하고 작업을 잃지 않으려면 계정을 생성하십시오.',
+        scan_qr_c2a: '다른 기기에서 이 세션으로 로그인하려면 아래 코드를 스캔하십시오',
+        select: "선택",
+        select_color: '색상 선택…',
+        send: "보내기",
+        send_password_recovery_email: "비밀번호 복구 이메일 보내기",
+        session_saved: "계정을 생성해 주셔서 감사합니다. 이 세션이 저장되었습니다.",
+        set_new_password: "새 비밀번호 설정",
+        share_to: "공유",
+        show_all_windows: "모든 창 표시",
+        show_hidden: '숨김 항목 표시',
+        sign_in_with_puter: "Puter로 로그인",
+        sign_up: "가입",
+        signing_in: "로그인 중…",
+        size: '크기',
+        sort_by: '정렬 기준',
+        start: '시작',
+        taking_longer_than_usual: '보통보다 조금 더 오래 걸립니다. 잠시만 기다려 주십시오...',
+        text_document: '텍스트 문서',
+        tos_fineprint: `무료 계정 생성을 클릭하면 Puter의 <a href="https://puter.com/terms" target="_blank">서비스 약관</a>과 <a href="https://puter.com/privacy" target="_blank">개인정보 보호정책</a>에 동의하는 것입니다.`,
+        trash: '휴지통',
+        type: '유형',
+        undo: '실행 취소',
+        unzip: "압축 해제",
+        upload: '업로드',
+        upload_here: '여기에 업로드',
+        username: "사용자 이름",
+        username_changed: '사용자 이름이 성공적으로 업데이트되었습니다.',
+        versions: "버전",
+        yes_release_it: '예, 해제합니다',
+        you_have_been_referred_to_puter_by_a_friend: "친구가 Puter로 추천했습니다!",
+        zip: "압축",
+    }
+};
+
+export default ko;
