+++ src/i18n/translations/translations.js
@@ -1,0 +1,15 @@
+import en from './en.js';
+import no from './no.js';
+import fa from './fa.js';
+import ko from './ko.js';
+import zh from './zh.js';
+import fr from './fr.js';
+
+export default {
+    en: en,
+    no: no,
+    fa: fa,
+    ko: ko,
+    zh: zh,
+    fr: fr,
+};
