+++ src/i18n/translations/zh.js
@@ -1,0 +1,148 @@
+const zh = {
+    name: "中文",
+    code: "zh",
+    dictionary: {
+        access_granted_to: "访问授权给",
+        add_existing_account: "添加现有帐户",
+        all_fields_required: '所有字段都是必需的。',
+        apply: "应用",
+        ascending: '升序',
+        background: "背景",
+        browse: "浏览",
+        cancel: '取消',
+        center: '中心',
+        change_desktop_background: '更改桌面背景…',
+        change_language: "更改语言",
+        change_password: "更改密码",
+        change_username: "更改用户名",
+        close_all_windows: "关闭所有窗口",
+        color: '颜色',
+        confirm_account_for_free_referral_storage_c2a: '创建帐户并确认您的电子邮件地址，以获得1 GB的免费存储空间。您的朋友也将获得1 GB的免费存储空间。',
+        confirm_new_password: "确认新密码",
+        contact_us: "联系我们",
+        contain: '包含',
+        continue: "继续",
+        copy: '复制',
+        copy_link: "复制链接",
+        copying: "复制",
+        cover: '封面',
+        create_account: "创建帐户",
+        create_free_account: "创建免费帐户",
+        create_shortcut: "创建快捷方式",
+        current_password: "当前密码",
+        cut: '剪切',
+        date_modified: '修改日期',
+        delete: '删除',
+        delete_permanently: "永久删除",
+        deploy_as_app: '部署为应用',
+        descending: '降序',
+        desktop_background_fit: "适合",
+        dir_published_as_website: `%strong% 已发布到：`,
+        disassociate_dir: "取消关联目录",
+        download: '下载',
+        downloading: "下载",
+        email: "电子邮件",
+        email_or_username: "电子邮件或用户名",
+        empty_trash: '清空回收站',
+        empty_trash_confirmation: `您确定要永久删除回收站中的项目吗？`,
+        emptying_trash: '清空回收站…',
+        feedback: "反馈",
+        feedback_c2a: "请使用下面的表格向我们发送您的反馈、评论和错误报告。",
+        feedback_sent_confirmation: "感谢您与我们联系。如果您的帐户关联有电子邮件，我们会尽快回复您。",
+        forgot_pass_c2a: "忘记密码？",
+        from: "从",
+        general: "一般",
+        get_a_copy_of_on_puter: `在 Puter.com 上获取 '%%' 的副本！`,
+        get_copy_link: '获取复制链接',
+        hide_all_windows: "隐藏所有窗口",
+        html_document: 'HTML 文档',
+        image: '图像',
+        invite_link: "邀请链接",
+        items_in_trash_cannot_be_renamed: `此项目无法重命名，因为它在回收站中。要重命名此项目，请先将其拖出回收站。`,
+        jpeg_image: 'JPEG 图像',
+        keep_in_taskbar: '保持在任务栏',
+        log_in: "登录",
+        log_out: '登出',
+        move: '移动',
+        moving: "移动",
+        my_websites: "我的网站",
+        name: '名称',
+        name_cannot_be_empty: '名称不能为空。',
+        name_cannot_contain_double_period: "名称不能是'..'字符。",
+        name_cannot_contain_period: "名称不能是'.'字符。",
+        name_cannot_contain_slash: "名称不能包含'/'字符。",
+        name_must_be_string: "名称只能是字符串。",
+        name_too_long: `名称不能超过 %% 个字符。`,
+        new: '新',
+        new_folder: '新文件夹',
+        new_password: "新密码",
+        new_username: "新用户名",
+        no_dir_associated_with_site: '此地址没有关联的目录。',
+        no_websites_published: "您尚未发布任何网站。",
+        ok: '好的',
+        open: "打开",
+        open_in_new_tab: "在新标签页中打开",
+        open_in_new_window: "在新窗口中打开",
+        open_with: "打开方式",
+        password: "密码",
+        password_changed: "密码已更改。",
+        passwords_do_not_match: '`新密码` 和 `确认新密码` 不匹配。',
+        paste: '粘贴',
+        paste_into_folder: "粘贴到文件夹",
+        pick_name_for_website: "为您的网站选择一个名称：",
+        picture: "图片",
+        powered_by_puter_js: `由 <a href="https://docs.puter.com/" target="_blank">Puter.js</a> 提供支持`,
+        preparing: "准备中...",
+        preparing_for_upload: "准备上传...",
+        properties: "属性",
+        publish: "发布",
+        publish_as_website: '发布为网站',
+        recent: "最近",
+        recover_password: "找回密码",
+        refer_friends_c2a: "每个创建并确认 Puter 帐户的朋友都会为您获得 1 GB。您的朋友也将获得 1 GB！",
+        refer_friends_social_media_c2a: `在 Puter.com 上获取 1 GB 的免费存储空间！`,
+        refresh: '刷新',
+        release_address_confirmation: `您确定要释放此地址吗？`,
+        remove_from_taskbar:'从任务栏中删除',
+        rename: '重命名',
+        repeat: '重复',
+        resend_confirmation_code: "重新发送确认码",
+        restore: "还原",
+        save_account_to_get_copy_link: "请创建帐户以继续。",
+        save_account_to_publish: '请创建帐户以继续。',
+        save_session_c2a: '创建帐户以保存当前会话，避免丢失工作。',
+        scan_qr_c2a: '扫描下面的代码以从其他设备登录此会话',
+        select: "选择",
+        select_color: '选择颜色…',
+        send: "发送",
+        send_password_recovery_email: "发送密码恢复电子邮件",
+        session_saved: "感谢您创建帐户。此会话已保存。",
+        set_new_password: "设置新密码",
+        share_to: "分享到",
+        show_all_windows: "显示所有窗口",
+        show_hidden: '显示隐藏',
+        sign_in_with_puter: "使用 Puter 登录",
+        sign_up: "注册",
+        signing_in: "登录中…",
+        size: '大小',
+        sort_by: '排序方式',
+        start: '开始',
+        taking_longer_than_usual: '需要的时间比平时长一点。请稍等...',
+        text_document: '文本文档',
+        tos_fineprint: `点击“创建免费帐户”即表示您同意 Puter 的 <a href="https://puter.com/terms" target="_blank">服务条款</a> 和 <a href="https://puter.com/privacy" target="_blank">隐私政策</a>。`,
+        trash: '回收站',
+        type: '类型',
+        undo: '撤销',
+        unzip: "解压缩",
+        upload: '上传',
+        upload_here: '在此上传',
+        username: "用户名",
+        username_changed: '用户名已成功更新。',
+        versions: "版本",
+        yes_release_it: '是的，释放它',
+        you_have_been_referred_to_puter_by_a_friend: "您已经被朋友推荐到 Puter！",
+        zip: "压缩",
+    }
+};
+
+export default zh;
