+++ src/initgui.js
@@ -47,8 +47,7 @@ window.initgui = async function(){
         puter.setAPIOrigin(api_origin);
 
     // determine locale
-    const userLang = navigator.language || navigator.userLanguage || 'en';
-    window.locale = userLang?.split('-')[0] ?? 'en';
+    window.locale = window.user_preferences.language;
 
     // Checks the type of device the user is on (phone, tablet, or desktop).
     // Depending on the device type, it sets a class attribute on the body tag
